Resolve circular dependencies, add missing arg check
