+++ src/crested/utils/__init__.py
@@ -2,12 +2,14 @@
 
 from ._logging import setup_logging
 from ._model_utils import permute_model
+from ._seq_utils import (
+    hot_encoding_to_sequence,
+    one_hot_encode_sequence,
+    reverse_complement,
+)
 from ._utils import (
     EnhancerOptimizer,
     extract_bigwig_values_per_bp,
     fetch_sequences,
-    hot_encoding_to_sequence,
-    one_hot_encode_sequence,
     read_bigwig_region,
-    reverse_complement,
 )
